Module-14-Hands-on-Implementation/03-gemini-demo/chat-with-pdf/requirements.txt
streamlit
PyMuPDF
google-generativeai
python-dotenv
